+++ packages/lineage/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @openparser/lineage
 
+## 1.0.4
+
 ## 1.0.3
 
 ## 1.0.2